Update requirements.txt
I've searched and understood better what the requirements.txt file is used for. However, the problem for me is linking the Github to the local Pycharm I downloaded. I've tried many ways to link them as I watched some videos about it but couldn't get the project to work locally.